+++ src/main/java/rdguez/eduardo/upax/model/response/EmployeePaymentResponse.java
@@ -5,13 +5,15 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 @Builder
 @Getter
 @Setter
 @ToString
 public class EmployeePaymentResponse {
 
-  private Double payment;
+  private BigDecimal payment;
 
   @Builder.Default
   private Boolean success = false;
